+++ src/SAD/Parser/FTL/Token.hs
@@ -13,9 +13,11 @@ module SAD.Parser.FTL.Token (
 ) where
 
 import Data.Text qualified as Text
+import Data.Text (Text)
 import FTLex.Ftl
 import Control.Monad.Trans.State.Strict (evalState, State)
 import Text.Megaparsec hiding (State, Token, token)
+import Control.Monad (unless)
 
 import SAD.Parser.Token
 import SAD.Parser.FTL.Lexer qualified as FTL
@@ -32,7 +34,7 @@ tokenize :: [FTL.Lexeme] -> IO [Token]
 tokenize lexemes = do
   filteredLexems <- filterLexemes lexemes
   case evalState (runParserT document "" filteredLexems) () of
-    Left err -> handleError (const unknownError) err
+    Left err -> handleError makeErrMsg err
     Right tokens -> return tokens
 
 -- | Take a list of lexemes, report all comments and remove all comments and
@@ -47,9 +49,21 @@ filterLexemes (l : ls) = case l of
   _ -> fmap (l :) (filterLexemes ls)
 
 
+-- * Tokenizing Errors and Warnings
+
+data Error = InvalidIsabelleSymbol Position.T Text
+  deriving (Eq, Ord)
+
+-- | Turn an error into a located error 
+makeErrMsg :: Error -> (Text, Position.T)
+makeErrMsg (InvalidIsabelleSymbol pos text) =
+  let msg = "Invalid Isabelle symbol " <> text <> "."
+  in (msg, pos)
+
+
 -- * Tokenizers
 
-type Tokenizer a = ParsecT () [FTL.Lexeme] (State ()) a
+type Tokenizer a = ParsecT Error [FTL.Lexeme] (State ()) a
 
 -- | Parse a whole FTL document.
 document :: Tokenizer [Token]
@@ -62,6 +76,7 @@ document = do
 token :: Tokenizer [Token]
 token = choice [
     symbol,
+    isabelleSymbol,
     word
   ]
 
@@ -71,6 +86,17 @@ symbol = do
   symbol <- satisfy isSymbolLexeme
   let text = Text.singleton $ symbolContent symbol
       pos = sourcePos symbol
+  return [Token text pos]
+
+-- | Parse a single symbol.
+isabelleSymbol :: Tokenizer [Token]
+isabelleSymbol = do
+  isabelleSymbol <- satisfy isIsabelleSymbolLexeme
+  let identifier = isabelleSymbolContent isabelleSymbol
+      text = "\\<" <> identifier <> ">"
+      pos = sourcePos isabelleSymbol
+  unless (identifier `elem` isabelleSymbols) $
+    customFailure (InvalidIsabelleSymbol pos text)
   return [Token text pos]
 
 -- | Parse a single word.